✅ Update: adjusting route files, configuring main route file and adjusting the application's main file syntax
